# src/main/resources/files-openapi.yaml
openapi: 3.0.1
info:
  title: Text File Processing API
  description: Text File Processing API
  version: 1.0.0
tags:
  - name: "files"
servers:
  - url: /v1
paths:
  /files/{taskId}:
    get:
      summary: Get information about a specific file by id
      operationId: getFileById
      tags:
        - "files"
      parameters:
        - name: taskId
          in: path
          required: true
          description: The id of a text file
          schema:
            type: string
      responses:
        '200':
          description: 'Expected response to a valid request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextFile'
        '404':
          description: 'File not found in the system'
          content:
            application/json:
              schema:
                type: string
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /files/ip/{ipAddress}:
    get:
      summary: Search for a file that contains a particular IP address
      operationId: getFileByIpAddress
      tags:
        - "files"
      parameters:
        - name: ipAddress
          in: path
          required: true
          description: An IP address contained by a text file
          schema:
            type: string
      responses:
        '200':
          description: 'Expected response to a valid request'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TextFile'
        '404':
          description: 'Specified IP address is not in the system'
          content:
            application/json:
              schema:
                type: string
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /files:
    post:
      summary: Create a new text file
      operationId: addTextFile
      tags:
        - "files"
      requestBody:
        description: Add a new text file into the system
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewFile"
        required: true
      responses:
        '201':
          description: A new text file was added successfully into the system
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Unable to add text file into the system
          content:
            application/json:
              schema:
                type: string
          x-codegen-request-body-name: user

components:
  schemas:
    NewFile:
      type: object
      properties:
        fileId:
          type: string
    TextFile:
      type: object
      properties:
        taskId:
          type: string
        taskCreationDate:
          type: string
          format: date
        fileId:
          type: string
        taskStatus:
          type: string
          enum:
            - created
            - completed
        ips:
          type: array
          items:
            type: string
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string